pyproject.toml
[tool.poetry]
name = "adaptive-tempscaling"
version = "0.1.0"
description = "Code for Adaptive TempScaling experiments"
authors = ["Sergio Alvarez Balanya"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^2.1.0"
torch = "^2.3"
matplotlib = "^3.9.2"
scipy = "^1.14.1"
torchvision = "^0.19.0"
scikit-learn = "^1.5.1"
pyro-ppl = "^1.9.1"
graphviz = "^0.20.3"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
ipython = "^8.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
